Cookie and session problem after many days solved
I have solved the cookie and login problem after many days. Now, the few data stored in the session, which are necessary for the frontend to work, have a limit of 1 day. Now if the user does not use the application in 24 hours, everything is reset thanks to the refresh cookie, but in case these expire, the frontend will not detect that the user is authenticated.